% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cub.R
\name{cub}
\alias{cub}
\title{Create a CUB observation model to estimate and model ordinal data.}
\usage{
cub(m)
}
\arguments{
\item{m}{Number of ordinal categories (responses are in 1, ..., m).}
}
\value{
A list containing:
\describe{
\item{probability}{Function to compute the probability of given responses under the CUB model.}
\item{log_probability}{Function to compute the log probability of given responses under the CUB model.}
\item{estimate_parameters}{Function to estimate the parameters of the CUB model.}
\item{number_of_parameters}{The number of parameters in the CUB model.}
}
}
\description{
The CUB model is a mixture of a deliberate choice
(shifted binomial with feeling parameter \code{xi}, \code{xi} close to 0 means high ratings)
and a random choice (uniform).
}
