% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cub.R
\name{cub_em}
\alias{cub_em}
\title{Perform the EM algorithm to estimate the parameters of a CUB model on a vector of responses}
\usage{
cub_em(r, m, n_iterations = 200, eps = 1e-06, bound = 0.001)
}
\arguments{
\item{r}{Vector of responses between 1 and m.}

\item{m}{Number of ordinal categories.}

\item{n_iterations}{Number of iterations for the EM algorithm.}

\item{eps}{Convergence threshold for the EM algorithm.}

\item{bound}{Keeps \code{pi_} and \code{xi} inside [bound, 1 - bound] to avoid log(0).}
}
\value{
A list with the estimated \code{pi_} and \code{xi}.
}
\description{
Perform the EM algorithm to estimate the parameters of a CUB model on a vector of responses
}
