% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cub.R
\name{cub_log_probability}
\alias{cub_log_probability}
\title{Calculate the log-probability of each cell of a block under the CUB model}
\usage{
cub_log_probability(block, m, pi_, xi)
}
\arguments{
\item{block}{Matrix of responses between 1 and m.}

\item{m}{Number of ordinal categories.}

\item{pi_}{Probability of a deliberate choice.}

\item{xi}{Feeling parameter.}
}
\value{
A block of log-probabilities corresponding to each response in \code{block}.
Missing responses (NA) give 0 so that they are ignored in a sum.
}
\description{
Calculate the log-probability of each cell of a block under the CUB model
}
