% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cub.R
\name{cub_probability}
\alias{cub_probability}
\title{Calculate the probability of a response under the CUB model}
\usage{
cub_probability(r, m, pi_, xi)
}
\arguments{
\item{r}{Vector of responses between 1 and m.}

\item{m}{Number of ordinal categories.}

\item{pi_}{Probability of a deliberate choice.}

\item{xi}{Feeling parameter.}
}
\value{
A vector of probabilities corresponding to each response in \code{r}.
}
\description{
Calculate the probability of a response under the CUB model
}
