% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cub.R
\name{e_zi__pi_xi}
\alias{e_zi__pi_xi}
\title{Calculate the posterior probability that each response comes from the
deliberate choice component rather than from the uniform component.}
\usage{
e_zi__pi_xi(r, m, pi_, xi)
}
\arguments{
\item{r}{Vector of responses between 1 and m (no NA).}

\item{m}{Number of ordinal categories.}

\item{pi_}{Current value of probability of deliberate choice.}

\item{xi}{Current value of feeling parameter.}
}
\value{
A vector tau with posterior probabilities corresponding to each response in \code{r}.
}
\description{
Calculate the posterior probability that each response comes from the
deliberate choice component rather than from the uniform component.
}
