% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{icl}
\alias{icl}
\title{Calculate the ICL of a model.}
\usage{
icl(model, ...)
}
\arguments{
\item{model}{A \code{model} object.}

\item{...}{Additional arguments.}
}
\value{
The ICL value.
}
\description{
Calculate the ICL of a model.
}
