% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbm.R
\name{initialize_lbm}
\alias{initialize_lbm}
\title{Initialize a Latent Block Model}
\usage{
initialize_lbm(x, K, L, model)
}
\arguments{
\item{x}{Matrix of observations.}

\item{K}{Number of row clusters.}

\item{L}{Number of column clusters.}

\item{model}{Observation model.}
}
\value{
An object of class \code{lbm}.
}
\description{
Initialize a Latent Block Model
}
