% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbm.R
\name{lbm}
\alias{lbm}
\title{Create a Latent Block Model}
\usage{
lbm(x, K, L, model, n_iterations, n_gibbs_iterations, eps, burn_in)
}
\arguments{
\item{x}{Matrix of observations.}

\item{K}{Number of row clusters.}

\item{L}{Number of column clusters.}

\item{model}{Observation model.}

\item{n_iterations}{Number of iterations for the SEM-Gibbs algorithm.}

\item{n_gibbs_iterations}{Number of Gibbs iterations for the SEM-Gibbs algorithm.}

\item{eps}{Convergence threshold for the SEM-Gibbs algorithm.}

\item{burn_in}{Number of burn-in iterations for the SEM-Gibbs algorithm.}
}
\value{
An object of class \code{lbm}.
}
\description{
Create a Latent Block Model
}
