% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Calculate the complete log-likelihood}
\usage{
log_likelihood(x, ...)
}
\arguments{
\item{x}{A \code{model} object.}

\item{...}{Additional arguments.}
}
\value{
The complete log-likelihood.
}
\description{
Calculate the complete log-likelihood
}
