% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocluster.R
\name{cocluster}
\alias{cocluster}
\title{Co-cluster data by fitting a latent block model using any observation model
and SEM-Gibbs estimation.}
\usage{
cocluster(
  x,
  K,
  L,
  model = cub,
  n_init = 20,
  n_iterations = 100,
  n_gibbs_iterations = 10,
  eps = 1e-09,
  burn_in = floor(n_iterations/2),
  choose_n_clusters = TRUE
)
}
\arguments{
\item{x}{Matrix of ordinal observations.}

\item{K}{Number of row clusters.}

\item{L}{Number of column clusters.}

\item{model}{Observation model.}

\item{n_init}{Number of random initializations.}

\item{n_iterations}{Number of iterations for the SEM-Gibbs algorithm.}

\item{n_gibbs_iterations}{Number of Gibbs iterations for the SEM-Gibbs algorithm.}

\item{eps}{Convergence threshold for the SEM-Gibbs algorithm.}

\item{burn_in}{Number of burn-in iterations for the SEM-Gibbs algorithm.}

\item{choose_n_clusters}{Whether the function will greedily select the best number of clusters based on ICL.}
}
\value{
An object of class \code{cocluster} containing:
\describe{
\item{lbms}{List of all models.}
\item{selected_lbm}{The selected model.}
\item{selected_lbm$icl}{ICL value of the selected model.}
\item{selected_lbm$p$row}{Estimated probabilities of row-cluster membership.}
\item{selected_lbm$p$col}{Estimated probabilities of column-cluster membership.}
\item{selected_lbm$z$row}{Final row-cluster assignments.}
\item{selected_lbm$z$col}{Final column-cluster assignments.}
\item{selected_lbm$block_parameters}{Estimated parameters of each block.}
}
}
\description{
Co-cluster data by fitting a latent block model using any observation model
and SEM-Gibbs estimation.
}
