% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cub.R
\name{cub}
\alias{cub}
\title{Create a CUB observation model}
\usage{
cub(m)
}
\arguments{
\item{m}{Number of ordinal categories (responses are in 1, ..., m).}
}
\value{
A list defining the CUB model.
}
\description{
Returns the list of functions expected by \code{\link[=cocluster]{cocluster()}}.
}
