% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbm.R
\name{icl.lbm}
\alias{icl.lbm}
\title{Calculate the ICL of the Latent Block Model}
\usage{
\method{icl}{lbm}(model)
}
\arguments{
\item{model}{An object of class \code{\link{lbm}}.}
}
\value{
The ICL value.
}
\description{
Calculate the ICL of the Latent Block Model
}
