% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lbm.R
\name{log_likelihood.lbm}
\alias{log_likelihood.lbm}
\title{Calculate the complete log-likelihood}
\usage{
\method{log_likelihood}{lbm}(lbm)
}
\arguments{
\item{lbm}{An \code{lbm} object.}
}
\value{
The complete log-likelihood.
}
\description{
Calculate the complete log-likelihood
}
