% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Calculate the complete log-likelihood of a model.}
\usage{
log_likelihood(model, ...)
}
\arguments{
\item{model}{A \code{model} object.}

\item{...}{Additional arguments.}
}
\value{
The complete log-likelihood.
}
\description{
Calculate the complete log-likelihood of a model.
}
