% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_likelihood}
\alias{log_likelihood}
\title{Calculate the complete log-likelihood of a model.}
\usage{
log_likelihood(model)
}
\arguments{
\item{model}{An object of class \code{model} containing:
\describe{
\item{x}{The data matrix.}
\item{z}{A list containing the cluster assignments.}
\item{p}{A list containing the prior probabilities of cluster membership.}
\item{block_parameters}{Estimated observation model parameters of each cluster.}
\item{model}{The observation model containing the \code{log_probability} function.}
}}
}
\value{
The complete log-likelihood.
}
\description{
Calculate the complete log-likelihood of a model.
}
